wip: structure reorganization
