+++ src/format/format/src/main.rs
@@ -1,307 +1,6 @@
 use serde::{self, Serialize, Deserialize};
 use std::collections::HashMap;
 
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Block {
-    block_type: BlockType,
-    geometry: Geometry,
-    id: String,
-    relationships: Vec<Relationship>,
-    confidence: f64,
-    text: String,
-    page: Option<i64>,
-    text_type: Option<TextType>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum BlockType {
-    #[serde(rename = "LINE")]
-    Line,
-    #[serde(rename = "PAGE")]
-    Page,
-    #[serde(rename = "WORD")]
-    Word,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Geometry {
-    bounding_box: BoundingBox,
-    polygon: Vec<Polygon>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct BoundingBox {
-    height: f64,
-    left: f64,
-    top: f64,
-    width: f64,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub struct Polygon {
-    x: f64,
-    y: f64,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Relationship {
-    ids: Vec<String>,
-    #[serde(rename = "Type")]
-    relationship_type: TypeEnum,
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum TypeEnum {
-    #[serde(rename = "CHILD")]
-    Child,
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum TextType {
-    #[serde(rename = "PRINTED")]
-    Printed,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StateResponseBlock {
-    confidence: f64,
-    text: String,
-}
-
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct LineItemGroup {
-    #[serde(rename = "LineItemGroupIndex")]
-    line_item_group_index: i64,
-    line_items: Vec<LineItem>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct LineItem {
-    line_item_expense_fields: Vec<LineItemExpenseField>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct LineItemExpenseField {
-    page_number: i64,
-    #[serde(rename = "Type")]
-    line_item_expense_field_type: TypeClass,
-    value_detection: Detection,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct TypeClass {
-    confidence: f64,
-    text: String,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Detection {
-    confidence: f64,
-    geometry: Geometry,
-    text: String,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StateResponseLineItemGroup {
-    line_item_group_index: i64,
-    line_items: Vec<StateResponseLineItem>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StateResponseLineItem {
-    line_item_expense_fields: Vec<StateResponseLineItemExpenseField>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub struct StateResponseLineItemExpenseField {
-    product_code: Option<StateResponseSummaryField>,
-    item: Option<StateResponseSummaryField>,
-    price: Option<StateResponseSummaryField>,
-    expense_row: Option<StateResponseSummaryField>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct StateResponseSummaryField {
-    confidence_key: f64,
-    confidence_value: f64,
-    value: String,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct SummaryFieldsWrapper {
-    summary_fields: Vec<SummaryField>,
-}
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct SummaryField {
-    group_properties: Option<Vec<GroupProperty>>,
-    page_number: i64,
-    #[serde(rename = "Type")]
-    summary_field_type: Option<TypeClass>,
-    value_detection: Detection,
-    label_detection: Option<Detection>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct GroupProperty {
-    id: String,
-    types: Vec<String>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StateResponseExpenseDocument {
-    blocks: Vec<StateResponseBlock>,
-    expense_index: i64,
-    line_item_groups: Vec<LineItemGroup>,
-    summary_fields: Vec<HashMap<String, StateResponseSummaryField>>,
-}
-
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct DocumentMetadata {
-    pages: i64,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StateResponse {
-    document_metadata: DocumentMetadata,
-    expense_documents: Vec<StateResponseExpenseDocument>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct ExpenseDocument {
-    blocks: Vec<Block>,
-    expense_index: i64,
-    line_item_groups: Vec<LineItemGroup>,
-    summary_fields: Vec<SummaryField>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct InputData{
-    blocks: Vec<Block>, 
-    line_item_groups: Vec<LineItemGroup>, 
-    summary_fields: [SummaryField]
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct OutputData{
-    blocks: Vec<StateResponseBlock>, 
-    line_item_groups: Vec<StateResponseLineItemGroup>, 
-    summary_fields:  Vec<HashMap<String, StateResponseSummaryField>>
-}
-
-
-// Function to process blocks
-fn process_blocks(blocks: &Vec<Block>) -> Vec<StateResponseBlock> {
-    blocks.iter().map(|block| {
-        StateResponseBlock{
-            confidence: block.confidence,
-            text: block.text.clone(),
-        }
-    }).collect()
-} 
-
-// Function to process LineItems
-fn process_line_item_groups(line_item_groups: &Vec<LineItemGroup>) -> Vec<StateResponseLineItemGroup> {
-    line_item_groups.iter().map(|line_item_group| {
-        StateResponseLineItemGroup {
-            line_item_group_index: line_item_group.line_item_group_index,
-            line_items: line_item_group.line_items.iter().map(|line_item| {
-                let line_item_expense_fields: Vec<StateResponseLineItemExpenseField> = line_item.line_item_expense_fields.iter().map(|field| {
-                    let mut expense_field = StateResponseLineItemExpenseField {
-                        product_code: None,
-                        item: None,
-                        expense_row: None,
-                        price: None
-                    };
-                    match field.line_item_expense_field_type.text.as_str()  {
-                        "PRODUCT_CODE" =>  expense_field.product_code = Some(StateResponseSummaryField {
-                            confidence_key: field.line_item_expense_field_type.confidence,
-                            confidence_value: field.value_detection.confidence,
-                            value: field.value_detection.text.clone(),
-                        }),
-                        "ITEM" =>  expense_field.item = Some(StateResponseSummaryField {
-                            confidence_key: field.line_item_expense_field_type.confidence,
-                            confidence_value: field.value_detection.confidence,
-                            value: field.value_detection.text.clone(),
-                        }),
-                        "EXPENSE_ROW" =>  expense_field.expense_row = Some(StateResponseSummaryField {
-                            confidence_key: field.line_item_expense_field_type.confidence,
-                            confidence_value: field.value_detection.confidence,
-                            value: field.value_detection.text.clone(),
-                        }),
-                        "PRICE" =>  expense_field.price = Some(StateResponseSummaryField {
-                            confidence_key: field.line_item_expense_field_type.confidence,
-                            confidence_value: field.value_detection.confidence,
-                            value: field.value_detection.text.clone(),
-                        }),
-                        _ => {} // Handling other cases if needed
-                    }
-                    expense_field
-                }).collect();
-                StateResponseLineItem { line_item_expense_fields }
-            }).collect()
-        }
-    }).collect()
-}
-
-
-fn process_summary_fields(summary_fields: &[SummaryField]) -> Vec<HashMap<String, StateResponseSummaryField>> {
-    summary_fields
-        .iter()
-        .map(|summary_field| {
-            let mut summary_map = HashMap::new();
-
-            let key = match summary_field.summary_field_type.as_ref().map(|t| t.text.as_str()) {
-                Some("OTHER") => summary_field.label_detection.as_ref().map_or("", |ld| ld.text.as_str()).to_string(),
-                _ => summary_field.summary_field_type.as_ref().map_or("", |t| t.text.as_str()).to_string(),
-            };
-
-            let confidence_key = match summary_field.summary_field_type {
-                Some(ref summary_field_type) => summary_field_type.confidence as f64, // Changed to f64
-                None => summary_field.label_detection.as_ref().map_or(0.0, |ld| ld.confidence as f64), // Changed to f64
-            };
-
-            summary_map.insert(
-                key,
-                StateResponseSummaryField {
-                    confidence_key,
-                    confidence_value: summary_field.value_detection.confidence,
-                    value: summary_field.value_detection.text.clone(),
-                },
-            );
-
-            summary_map
-        })
-        .collect()
-}
-
-// Handler function
 fn function_handler(data: InputData) {
     let processed_blocks = process_blocks(data.blocks);
 
@@ -320,7 +19,6 @@ fn function_handler(data: InputData) {
 fn main(data: InputData){
     function_handler(data)
 }
-
 
 #[cfg(test)]
 mod tests {
