+++ src/format/format/src/type.rs
@@ -1,0 +1,6 @@
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct TypeClass {
+    confidence: f64,
+    text: String,
+}
